+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Lagt til select-komponent
 - Lagt til typografi
+- Lagt til toast-komponent
+- Endret verdien på propen "prioritet" på alert-komponenten fra 'Info' | 'Danger' | 'Warning' | 'Success' til 'info' | 'danger' | 'warning' | 'success'
+- Fjernet default header verdi på alert-komponenten.
 
 # 0.0.8 - 13.01.2025